reorganise components
