+++ src/components/DetailPage/CommentsBlock/Comments.tsx
@@ -9,10 +9,10 @@ import { Publication, PublicationsDocument } from '@/types/lens';
 import { useRouter } from 'next/router';
 import { sanitizeIpfsUrl } from '@/utils/sanitizeIpfsUrl';
 import CreateComment from './CreateComment';
-import LoginButton from './LoginButton';
+import LoginButton from '../../LoginButton';
 import getAvatar from '@/lib/getAvatar';
 import CommentData from './CommentData';
-import QueuedData from './QueuedData';
+import QueuedData from '../../QueuedData';
 
 interface Props {
     publication: Publication;
