+++ src/components/DetailPage/CommentsBlock/NoResults.tsx
