+++ src/components/DetailPage/DetailRender.tsx
@@ -5,10 +5,10 @@ import { useAppPersistStore, useAppStore, useReferenceModuleStore } from "@/stor
 import { useAccount, useDisconnect, useNetwork } from 'wagmi';
 import { Profile, ReferenceModules, UserProfilesDocument, UserProfilesQuery, UserProfilesQueryVariables } from "@/types/lens";
 import { CHAIN_ID } from "@/constants";
-import Loading from "./Loading";
-import ProfileCard from './ProfileCard';
-import Profiles from './Profiles';
-import VideoDetail from './UI/VideoDetail';
+import Loading from "../Loading";
+import ProfileCard from '../ProfilePage/ProfileCard';
+import Profiles from '../ProfilePage/Profiles';
+import VideoDetail from './VideoDetail';
 
 
 const ProfileRender = () => {
