+++ src/components/DetailPage/VideoDetail.tsx
@@ -8,7 +8,7 @@ import { MdOutlineCancel } from 'react-icons/md';
 import { BsFillPlayFill } from 'react-icons/bs';
 import { HiVolumeUp, HiVolumeOff } from 'react-icons/hi';
 
-import Comments from '../Comments';
+import Comments from './CommentsBlock/Comments';
 import { useQuery } from '@apollo/client';
 import { Publication, PublicationDocument, Profile } from '@/types/lens';
 import getMedia from '@/lib/getMedia';
@@ -18,10 +18,10 @@ import { copyToClipboard } from "@/utils/clipboard";
 import { AiFillHeart, AiFillTwitterCircle } from "react-icons/ai";
 import { BsFacebook, BsReddit } from "react-icons/bs";
 import { FaCommentDots, FaTimes } from "react-icons/fa";
-import LikeButton from '../Buttons/LikeButton';
-import MirrorButton from '../Buttons/MirrorButton';
+import LikeButton from '../Buttons/Likes/LikeButton';
+import MirrorButton from '../Buttons/Mirrors/MirrorButton';
 import CommentButton from '../Buttons/CommentButton';
-import CreateComment from '../CreateComment';
+import CreateComment from './CommentsBlock/CreateComment';
 import LoginButton from '../LoginButton';
 import { useAppStore } from "src/store/app";
 import UnfollowButton from '../Buttons/UnfollowButton';
