+++ src/pages/detail/[id].tsx
@@ -1,3 +1,3 @@
-import DetailRender from "@/components/DetailRender";
+import DetailRender from "@/components/DetailPage/DetailRender";
 
 export default DetailRender
